Frozen modal back in with new modal concept
